+++ components/design/phoneInput/index.ts
@@ -1,0 +1,4 @@
+export { default as PhoneInput, getPhoneValue, PHONE_COUNTRIES } from "./PhoneInput";
+export { countryFlagUrl, preferredCountries } from "./countries";
+export type { PhoneCountry, PhoneFormat, PhoneInputProps, PhoneValue } from "./types";
+export { default } from "./PhoneInput";